+++ client/src/components/Forms/index.js
@@ -1,3 +1,3 @@
 export { LoginForm } from "./LoginForm";
 export { SignupForm } from './SignupForm'
-export { AddStudentForm } from "./AddStudentForm";
+export { FormInput } from './FormInput'
